Added a stochastic haplotype enumerator for complex regions with too many haplotypes to analyze

+++ src/main/java/org/broadinstitute/hellbender/tools/walkers/haplotypecaller/graphs/KBestSubHaplotypeFinder.java
@@ -38,7 +38,7 @@ interface KBestSubHaplotypeFinder {
      * Returns the total number of possible sub-haplotypes.
      * @return 0 or greater.
      */
-    public int getCount();
+    public long getCount();
 
     /**
      * Return the k-best sub-haplotype solution.
